Memory_Me_Please/requirements.txt
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
rich==13.7.0
backoff==2.2.1
typing-extensions==4.9.0
